% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqseq.R
\name{qqseq}
\alias{qqseq}
\title{Title}
\usage{
qqseq(transcript_id, species, max_retries = 20)
}
\arguments{
\item{transcript_id}{esembl transcript_id}

\item{species}{species}

\item{max_retries}{he maximum number of retries allowed. Default is 20.}
}
\value{
sequences of transcript
}
\description{
Title
}
\examples{
result <- qqseq("ENSMUST00000127786", "mouse")
}
